fetch-data/requirements.txt
Flask
python-dotenv
requests
mysql-connector-python
functions-framework
defusedxml